Refactor, tweak how token is made ... slightly
