+++ betterself/users/models.py
@@ -71,5 +71,4 @@ def post_delete_user(sender, instance, *args, **kwargs):
 
 @receiver(post_save, sender=settings.AUTH_USER_MODEL)
 def create_auth_token(sender, instance=None, created=False, **kwargs):
-    if created:
-        Token.objects.create(user=instance)
+    Token.objects.get_or_create(user=instance)
