+++ config/settings/settings_testing.py
@@ -81,3 +81,5 @@ LOGGING = {
 # now celery tasks autorun -- so ... write unit tests for your celery tasks.
 # TODO - Do this and do it well
 # CELERY_TASK_ALWAYS_EAGER = True
+
+MIGRATION_MODULES = {}
